requirements.txt
anthropic>=0.40.0
slack-bolt>=1.20.0
slack-sdk>=3.30.0
notion-client>=2.2.0
python-dotenv>=1.0.0
